deprecated: `base_validation_wrapper` is moved to `BaseValidationParameters`
Having a decorator `base_validation_wrapper` was proving to be
completely useless as it could of been declared in
`BaseValidationParameters`, hence, moving all the function to the super
class.
