+++ validoopsie/base/__init__.py
@@ -1,7 +1,5 @@
 from .base_validation_parameters import BaseValidationParameters
-from .base_validation_wrapper import base_validation_wrapper
 
 __all__ = [
     "BaseValidationParameters",
-    "base_validation_wrapper",
 ]
